requirements.txt
requests
python-dotenv
pyyaml
openai
playsound
gTTS
git+https://github.com/openai/whisper.git
pydub
PyAudio
playsound